reformatting the repo to be pythonic
